[SUN-44] refactor: changing the project structure
