+++ src/main/java/com/jovicsantos/suneverapi/api/controller/IngredientController.java
@@ -1,4 +1,4 @@
-package com.jovicsantos.suneverapi.controllers;
+package com.jovicsantos.suneverapi.api.controller;
 
 import java.util.UUID;
 
@@ -15,10 +15,10 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.jovicsantos.suneverapi.dtos.IngredientDto;
-import com.jovicsantos.suneverapi.models.Ingredient;
-import com.jovicsantos.suneverapi.services.IngredientService;
-import com.jovicsantos.suneverapi.services.MeasurementService;
+import com.jovicsantos.suneverapi.application.dto.IngredientDto;
+import com.jovicsantos.suneverapi.domain.entity.Ingredient;
+import com.jovicsantos.suneverapi.infrastructure.service.IngredientService;
+import com.jovicsantos.suneverapi.infrastructure.service.MeasurementService;
 
 import jakarta.validation.Valid;
 import lombok.var;
