+++ src/main/java/com/jovicsantos/suneverapi/application/dto/MeasurementDto.java
@@ -1,4 +1,4 @@
-package com.jovicsantos.suneverapi.dtos;
+package com.jovicsantos.suneverapi.application.dto;
 
 import org.hibernate.validator.constraints.Length;
