+++ src/main/java/com/jovicsantos/suneverapi/application/dto/RecipeDto.java
@@ -1,7 +1,7 @@
-package com.jovicsantos.suneverapi.dtos;
+package com.jovicsantos.suneverapi.application.dto;
 
 import java.util.List;
 
 public record RecipeDto(String name, String description, String imageLink, Integer portions,
-    List<RecipeIngredientsDto> ingredients) {
+                List<RecipeIngredientsDto> ingredients) {
 }
