+++ src/main/java/com/jovicsantos/suneverapi/domain/entity/Ingredient.java
@@ -1,12 +1,9 @@
-package com.jovicsantos.suneverapi.models;
+package com.jovicsantos.suneverapi.domain.entity;
 
 import java.io.Serializable;
 import java.math.BigDecimal;
 import java.util.UUID;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
-
-import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -19,23 +16,21 @@ import lombok.Data;
 
 @Entity
 @Data
-@Table(name = "RecipeIngredient")
-public class RecipeIngredient implements Serializable {
+@Table(name = "ingredient")
+public class Ingredient implements Serializable {
   private static final long serialVersionUID = 1L;
 
   @Id
   @GeneratedValue(strategy = GenerationType.AUTO)
-  @JsonIgnore
-  private UUID recipeIngredientId;
-
-  @JsonIgnore
-  @ManyToOne(cascade = CascadeType.ALL)
-  @JoinColumn(name = "recipeId")
-  private Recipe recipe;
-
-  @Column(nullable = false, name = "ingredient_id")
   private UUID id;
+  @Column(nullable = false)
+  private String name;
+  @Column(nullable = false, precision = 10, scale = 2)
+  private BigDecimal price;
+  @Column(nullable = false, precision = 10, scale = 2)
+  private BigDecimal quantityPerMeasure;
 
-  @Column(nullable = false, name = "ingredient_quantity")
-  private BigDecimal quantity;
+  @ManyToOne
+  @JoinColumn(name = "measurement_id")
+  private Measurement measurement;
 }
