+++ src/main/java/com/jovicsantos/suneverapi/infrastructure/repository/IngredientRepository.java
@@ -1,10 +1,10 @@
-package com.jovicsantos.suneverapi.repositories;
+package com.jovicsantos.suneverapi.infrastructure.repository;
 
 import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.jovicsantos.suneverapi.models.Ingredient;
+import com.jovicsantos.suneverapi.domain.entity.Ingredient;
 
 public interface IngredientRepository extends JpaRepository<Ingredient, UUID> {
   boolean existsByName(String name);
