+++ src/main/java/com/jovicsantos/suneverapi/infrastructure/repository/MeasurementRepository.java
@@ -1,10 +1,10 @@
-package com.jovicsantos.suneverapi.repositories;
+package com.jovicsantos.suneverapi.infrastructure.repository;
 
 import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.jovicsantos.suneverapi.models.Measurement;
+import com.jovicsantos.suneverapi.domain.entity.Measurement;
 
 public interface MeasurementRepository extends JpaRepository<Measurement, UUID> {
   boolean existsByName(String name);
