+++ src/main/java/com/jovicsantos/suneverapi/infrastructure/repository/RecipeIngredientRepository.java
@@ -1,10 +1,10 @@
-package com.jovicsantos.suneverapi.repositories;
+package com.jovicsantos.suneverapi.infrastructure.repository;
 
 import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.jovicsantos.suneverapi.models.RecipeIngredient;
+import com.jovicsantos.suneverapi.domain.entity.RecipeIngredient;
 
 public interface RecipeIngredientRepository extends JpaRepository<RecipeIngredient, UUID> {
