+++ src/main/java/com/jovicsantos/suneverapi/infrastructure/repository/RecipeRepository.java
@@ -1,10 +1,10 @@
-package com.jovicsantos.suneverapi.repositories;
+package com.jovicsantos.suneverapi.infrastructure.repository;
 
 import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.jovicsantos.suneverapi.models.Recipe;
+import com.jovicsantos.suneverapi.domain.entity.Recipe;
 
 public interface RecipeRepository extends JpaRepository<Recipe, UUID> {
   boolean existsByName(String name);
